Crate zink

Zink library for developing smart contracts for blockchains.

Re-exports

pub use storage::DoubleKeyMapping;

pub use storage::Mapping;

pub use storage::Storage;

pub use storage::TransientStorage;

Modules

asm 🔒

Assembly trait implementation.

event 🔒

ffi

Zink FFI.

primitives

Zink primitive types

storage

Zink storage implementation.

Macros

assert

Check and expression and revert with the input message

import

The import! macro generates a Rust struct and implementation for interacting with an Ethereum smart contract based on its ABI (Application Binary Interface) and deploys the corresponding contract.

revert

Revert with the input message

Traits

Asm

Types implemented this trait are able to be pushed on stack.

Event

Zink event interface

Functions

keccak256

Generate a keccak hash of the input (sha3)

to_bytes32

Convert bytes to ls bytes

Attribute Macros

external

Mark the function as an external entry point.

storage

Declare on-chain storage

transient_storage

Declare transient storage (cleared after each transaction)

Derive Macros

Event

Event logging interface

Storage

Declare contract storage