zink::primitives

Module numeric

Macros

impl_numeric 🔒

impl_safe_numeric_signed 🔒

impl_safe_numeric_unsigned 🔒

local_revert 🔒

Traits

Numeric

A trait for modular arithmetic operations on numeric types.

SafeNumeric

A trait for safe arithmetic operations with bound checks.